+++ CHANGELOG.md
@@ -23,6 +23,7 @@
    - *this.update -> this._update*
    - _initScene()
 - #53: Wagner moved to plugins.
+- Fixed resize. *rWidth, rHeight default parameter is now 1.*
 
 **v0.0.7**
 - Fixed #16 "object.assign() chrome bug."